[TASK2] Удаление ненужных интерфейсов + исправление комментариев + добавление хендлеров
