+++ src/main/java/ru/virtya/vacationPayCalculate/dto/ErrorDto.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 
-@AllArgsConstructor
 @Getter
 @Setter
+@AllArgsConstructor
 public class ErrorDto {
 
     private String message;
