+++ src/main/java/ru/virtya/vacationPayCalculate/handler/GlobalControllerExceptionHandler.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import ru.virtya.vacationPayCalculate.dto.ErrorDto;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
 
 @ControllerAdvice
 public class GlobalControllerExceptionHandler {
@@ -18,6 +19,22 @@ public class GlobalControllerExceptionHandler {
 
         return new ResponseEntity<>(
                 new ErrorDto(e.getMessage(), LocalDate.now()), HttpStatus.BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public @ResponseBody ResponseEntity<ErrorDto> handleIllegalArgument(IllegalArgumentException e) {
+
+        return new ResponseEntity<>(
+                new ErrorDto(e.getMessage(), LocalDate.now()), HttpStatus.INTERNAL_SERVER_ERROR
+        );
+    }
+
+    @ExceptionHandler(DateTimeParseException.class)
+    public @ResponseBody ResponseEntity<ErrorDto> handleDateTimeParse(DateTimeParseException e) {
+
+        return new ResponseEntity<>(
+                new ErrorDto(e.getMessage(), LocalDate.now()), HttpStatus.INTERNAL_SERVER_ERROR
         );
     }
 }
